fixes in rooms module

+++ NarwianskiZakatek/Controllers/RoomsController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using NarwianskiZakatek.Data;
 using NarwianskiZakatek.Models;
+using NarwianskiZakatek.Services;
 using System.Data;
 
 namespace NarwianskiZakatek.Controllers
@@ -10,19 +11,22 @@ namespace NarwianskiZakatek.Controllers
     public class RoomsController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly IEmailService _sender;
 
-        public RoomsController(ApplicationDbContext context)
+        public RoomsController(ApplicationDbContext context, IEmailService sender)
         {
             _context = context;
+            _sender = sender;
         }
 
         // GET: Rooms
         [Authorize(Roles = "Admin,Employee")]
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string? message)
         {
-              return _context.Rooms != null ? 
-                          View(await _context.Rooms.ToListAsync()) :
-                          Problem("Entity set 'ApplicationDbContext.Rooms'  is null.");
+            ViewBag.Message = message;
+            return _context.Rooms != null ?
+                        View(await _context.Rooms.OrderBy(r => r.RoomNumber).ToListAsync()) :
+                        Problem("Entity set 'ApplicationDbContext.Rooms'  is null.");
         }
 
         // GET: Rooms/Create
@@ -82,8 +86,22 @@ namespace NarwianskiZakatek.Controllers
 
             if (ModelState.IsValid)
             {
+                string message = "Pokój został zaktualizowany.";
                 try
                 {
+                    var oldRoom = _context.Rooms.Where(r => r.RoomId == id).AsNoTracking().First();
+                    if(oldRoom.IsActive && !room.IsActive)
+                    {
+                        var reservationIds = _context.ReservedRooms.Where(rr => rr.RoomId == oldRoom.RoomId).Select(rr => rr.ReservationId).ToList();
+                        var reservations = _context.Reservations.Where(r => reservationIds.Contains(r.ReservationId)).Include(r => r.User).ToList();
+                        for (int i = 0; i < reservations.Count; i++)
+                        {
+                            reservations[i].IsCancelled = true;
+                            _sender.CancelReservationAsync(reservations[i].User.Email, reservations[i]);
+                        }
+                        message = "Pokój został zaktualizowany. Pokój został wyłączony z użycia, a wszystkie nadchodzące rezerwacje tego pokoju zostały odwołane.";
+                    }
+                    
                     _context.Update(room);
                     await _context.SaveChangesAsync();
                 }
@@ -98,7 +116,10 @@ namespace NarwianskiZakatek.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index", new
+                {
+                    message = message
+                });
             }
             return View(room);
         }
@@ -128,23 +149,30 @@ namespace NarwianskiZakatek.Controllers
         [Authorize(Roles = "Admin,Employee")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            string message = "Pokój nie został usunięty, gdyż istnieją w systemie powiązane z nim rezerwacje." +
+                " Jeżeli chcesz wycofać pokój z użytku, zmień jego dostępność w oknie edycji pokoju.";
             if (_context.Rooms == null)
             {
                 return Problem("Entity set 'ApplicationDbContext.Rooms'  is null.");
             }
             var room = await _context.Rooms.FindAsync(id);
-            if (room != null)
+            bool canDelete = room?.ReservedRooms?.Count == 0;
+            if (room != null && canDelete)
             {
                 _context.Rooms.Remove(room);
+                message = "Pomyślnie usunięto pokój o numerze " + room.RoomNumber + ".";
             }
-            
+
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Index", new
+            {
+                message = message
+            });
         }
 
         private bool RoomExists(int id)
         {
-          return (_context.Rooms?.Any(e => e.RoomId == id)).GetValueOrDefault();
+            return (_context.Rooms?.Any(e => e.RoomId == id)).GetValueOrDefault();
         }
     }
 }
